pyproject.toml
[tool.poetry]
name = "themis-backend"
version = "0.1.0"
description = ""
authors = ["Gabriel Soares Baptista <gsoaresbaptista@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*,<4.0"
starlette = "^0.32.0.post1"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0.post1"

[tool.taskipy.tasks]
format = 'blue .  && isort .'
run = "uvicorn themis_backend.application.app:app --reload"
lint = 'ruff . && blue --check . --diff'

[tool.ruff]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.0"
blue = "^0.9.1"
ruff = "^0.1.5"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
